Made a lot of changes to clean up the code. Handled similar answers. Fixed logic for adding cards. Deleted unused files. Split the gui.py file into two separate files: gui.py and main.py. gui.py has everything related to the GUI windows, while the main.py has the event loop. Deleted everything related to the prototype.
