--- Instructions.txt
@@ -1,12 +1,0 @@
-1. Divide the class into two teams.
-2. One member of each team faces the other in a face-off as the teacher reads the question off the game board.
-3. The team that buzzed in with the correct answer receives control of the board and has the option of playing or passing control to the other team.
-4. The team that has the control tries to reveal all of the correct answers to the question before receiving three strikes. (If the answer is not on the board, click the Try Again button.)
-5. If the team receives three strikes without clearing the board, control is passed to the other team.
-6. The team that now has the control is able to give one answer in the hopes that it is found on the board.
-	o If it is, points are added to the team's score.
-	o If not, the other team gets the points.
-7. Points are collected as each team finds its answers to the question on the board. (Click the Score button to add points to the appropriate team's score.)
-8. Continue the game repeating steps two through seven.
-9. Each team tries to collect the most points. The team with the most points wins the game! (Click on Team 1 or Team 2 on the Score slide to move to the You've Won slide.)
-
